only owner and admin can update delete the bootcamp

+++ controllers/bootcamps.js
@@ -60,14 +60,21 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 // @access  Public
 
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
-    const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
+    let bootcamp = await Bootcamp.findById(req.params.id)
+
+    if (!bootcamp) {
+        return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404))
+    }
+
+    // make sure user is bootcamp owner 
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`User ${req.params.id} not authorized to update this bootcamp`, 401))
+    }
+
+    bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
         runValidators: true
     })
-
-    if (!bootcamp) {
-        return res.status(400).json({ success: false })
-    }
 
     res.status(200).json({ success: true, data: bootcamp })
 })
@@ -81,6 +88,12 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
     if (!bootcamp) {
         return next(new ErrorResponse(`No bootcamp of id ${req.params.id}`, 400))
     }
+
+    // make sure user is bootcamp owner 
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`User ${req.params.id} not authorized to delete this bootcamp`, 401))
+    }
+
     // delete course 
     await bootcamp.deleteOne({ _id: req.params.id })
 
@@ -128,6 +141,11 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
 
     if (!bootcamp) {
         return next(new ErrorResponse(`Bootcamp not found with the id of ${req.param.id}`, 404))
+    }
+
+    // make sure user is bootcamp owner 
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`User ${req.params.id} not authorized to update this bootcamp`, 401))
     }
 
     if (!req.files) {
